Split instance related handlers into dimensions and import for clarity
