+++ instance/instance.go
@@ -229,101 +229,6 @@ func (s *Store) Add(w http.ResponseWriter, r *http.Request) {
 	log.InfoCtx(ctx, "add instance: request successful", logData)
 }
 
-// UpdateDimension updates label and/or description
-// for a specific dimension within an instance
-func (s *Store) UpdateDimension(w http.ResponseWriter, r *http.Request) {
-
-	defer request.DrainBody(r)
-
-	ctx := r.Context()
-	vars := mux.Vars(r)
-	instanceID := vars["instance_id"]
-	dimension := vars["dimension"]
-	auditParams := common.Params{"instance_id": instanceID, "dimension": dimension}
-	logData := audit.ToLogData(auditParams)
-
-	log.InfoCtx(ctx, "update instance dimension", logData)
-
-	if err := func() error {
-		instance, err := s.GetInstance(instanceID)
-		if err != nil {
-			log.ErrorCtx(ctx, errors.WithMessage(err, "update instance dimension: Failed to GET instance"), logData)
-			return err
-		}
-		auditParams["instance_state"] = instance.State
-
-		// Early return if instance state is invalid
-		if err = models.CheckState("instance", instance.State); err != nil {
-			logData["state"] = instance.State
-			log.ErrorCtx(ctx, errors.WithMessage(err, "update instance dimension: current instance has an invalid state"), logData)
-			return err
-		}
-
-		// Read and unmarshal request body
-		b, err := ioutil.ReadAll(r.Body)
-		if err != nil {
-			log.ErrorCtx(ctx, errors.WithMessage(err, "update instance dimension: error reading request.body"), logData)
-			return errs.ErrUnableToReadMessage
-		}
-
-		var dim *models.Dimension
-
-		err = json.Unmarshal(b, &dim)
-		if err != nil {
-			log.ErrorCtx(ctx, errors.WithMessage(err, "update instance dimension: failing to model models.Codelist resource based on request"), logData)
-			return errs.ErrUnableToParseJSON
-		}
-
-		// Update instance-dimension
-		notFound := true
-		for i := range instance.Dimensions {
-
-			// For the chosen dimension
-			if instance.Dimensions[i].Name == dimension {
-				notFound = false
-				// Assign update info, conditionals to allow updating
-				// of both or either without blanking other
-				if dim.Label != "" {
-					instance.Dimensions[i].Label = dim.Label
-				}
-				if dim.Description != "" {
-					instance.Dimensions[i].Description = dim.Description
-				}
-				break
-			}
-		}
-
-		if notFound {
-			log.ErrorCtx(ctx, errors.WithMessage(errs.ErrDimensionNotFound, "update instance dimension: dimension not found"), logData)
-			return errs.ErrDimensionNotFound
-		}
-
-		// Only update dimensions of an instance
-		instanceUpdate := &models.Instance{
-			Dimensions:      instance.Dimensions,
-			UniqueTimestamp: instance.UniqueTimestamp,
-		}
-
-		// Update instance
-		if err = s.UpdateInstance(ctx, instanceID, instanceUpdate); err != nil {
-			log.ErrorCtx(ctx, errors.WithMessage(err, "update instance dimension: failed to update instance with new dimension label/description"), logData)
-			return err
-		}
-
-		return nil
-	}(); err != nil {
-		if auditErr := s.Auditor.Record(ctx, UpdateDimensionAction, audit.Unsuccessful, auditParams); auditErr != nil {
-			err = auditErr
-		}
-		handleInstanceErr(ctx, err, w, logData)
-		return
-	}
-
-	s.Auditor.Record(ctx, UpdateDimensionAction, audit.Successful, auditParams)
-
-	log.InfoCtx(ctx, "updated instance dimension: request successful", logData)
-}
-
 //Update a specific instance
 func (s *Store) Update(w http.ResponseWriter, r *http.Request) {
 
@@ -531,273 +436,6 @@ func validateInstanceStateUpdate(instance, currentInstance *models.Instance) (er
 	}
 
 	return nil
-}
-
-func (s *Store) getEdition(ctx context.Context, datasetID, edition, instanceID string) (*models.EditionUpdate, string, error) {
-	auditParams := common.Params{"dataset_id": datasetID, "instance_id": instanceID, "edition": edition}
-	logData := audit.ToLogData(auditParams)
-
-	var action string
-
-	editionDoc, err := s.GetEdition(datasetID, edition, "")
-	if err != nil {
-		if err != errs.ErrEditionNotFound {
-			log.ErrorCtx(ctx, err, logData)
-			return nil, action, err
-		}
-
-		action = CreateEditionAction
-		if auditErr := s.Auditor.Record(ctx, action, audit.Attempted, auditParams); auditErr != nil {
-			return nil, action, auditErr
-		}
-
-		// create unique id for edition
-		editionID := uuid.NewV4().String()
-
-		editionDoc = &models.EditionUpdate{
-			ID: editionID,
-			Next: &models.Edition{
-				Edition: edition,
-				State:   models.EditionConfirmedState,
-				Links: &models.EditionUpdateLinks{
-					Dataset: &models.LinkObject{
-						ID:   datasetID,
-						HRef: fmt.Sprintf("%s/datasets/%s", s.Host, datasetID),
-					},
-					Self: &models.LinkObject{
-						HRef: fmt.Sprintf("%s/datasets/%s/editions/%s", s.Host, datasetID, edition),
-					},
-					Versions: &models.LinkObject{
-						HRef: fmt.Sprintf("%s/datasets/%s/editions/%s/versions", s.Host, datasetID, edition),
-					},
-					LatestVersion: &models.LinkObject{
-						ID:   "1",
-						HRef: fmt.Sprintf("%s/datasets/%s/editions/%s/versions/1", s.Host, datasetID, edition),
-					},
-				},
-			},
-		}
-	} else {
-		action = UpdateEditionAction
-		if auditErr := s.Auditor.Record(ctx, action, audit.Attempted, auditParams); auditErr != nil {
-			return nil, action, auditErr
-		}
-
-		// Update the latest version for the dataset edition
-		version, err := strconv.Atoi(editionDoc.Next.Links.LatestVersion.ID)
-		if err != nil {
-			logData["version"] = editionDoc.Next.Links.LatestVersion.ID
-			log.ErrorCtx(ctx, errors.WithMessage(err, "unable to retrieve latest version"), logData)
-			return nil, UpdateEditionAction, err
-		}
-
-		version++
-
-		editionDoc.Next.Links.LatestVersion.ID = strconv.Itoa(version)
-		editionDoc.Next.Links.LatestVersion.HRef = fmt.Sprintf("%s/datasets/%s/editions/%s/versions/%s", s.Host, datasetID, edition, strconv.Itoa(version))
-	}
-
-	return editionDoc, action, nil
-}
-
-func (s *Store) addInstanceLinks(ctx context.Context, instance *models.Instance, editionDoc *models.EditionUpdate) *models.Instance {
-	stringifiedVersion := strconv.Itoa(instance.Version)
-
-	log.InfoCtx(ctx, "adding instance links", log.Data{"editionDoc": editionDoc.Next, "instance": instance})
-
-	if instance.Links == nil {
-		instance.Links = &models.InstanceLinks{}
-	}
-
-	instance.Links.Dataset = &models.IDLink{
-		HRef: editionDoc.Next.Links.Dataset.HRef,
-		ID:   editionDoc.Next.Links.Dataset.ID,
-	}
-
-	instance.Links.Dimensions = &models.IDLink{
-		HRef: fmt.Sprintf("%s/versions/%s/dimensions", editionDoc.Next.Links.Self.HRef, stringifiedVersion),
-	}
-
-	instance.Links.Edition = &models.IDLink{
-		HRef: editionDoc.Next.Links.Self.HRef,
-		ID:   editionDoc.Next.Edition,
-	}
-
-	instance.Links.Version = &models.IDLink{
-		HRef: fmt.Sprintf("%s/versions/%s", editionDoc.Next.Links.Self.HRef, stringifiedVersion),
-		ID:   stringifiedVersion,
-	}
-
-	return instance
-}
-
-// UpdateObservations increments the count of inserted_observations against
-// an instance
-func (s *Store) UpdateObservations(w http.ResponseWriter, r *http.Request) {
-
-	ctx := r.Context()
-	vars := mux.Vars(r)
-	instanceID := vars["instance_id"]
-	insert := vars["inserted_observations"]
-	auditParams := common.Params{"instance_id": instanceID, "inserted_observations": insert}
-	logData := audit.ToLogData(auditParams)
-
-	if err := func() error {
-		observations, err := strconv.ParseInt(insert, 10, 64)
-		if err != nil {
-			log.ErrorCtx(ctx, errors.WithMessage(err, "update imported observations: failed to parse inserted_observations string to int"), logData)
-			return errs.ErrInsertedObservationsInvalidSyntax
-		}
-
-		if err = s.UpdateObservationInserted(instanceID, observations); err != nil {
-			log.ErrorCtx(ctx, errors.WithMessage(err, "update imported observations: store.UpdateObservationInserted returned an error"), logData)
-			return err
-		}
-
-		return nil
-	}(); err != nil {
-		if auditErr := s.Auditor.Record(ctx, UpdateInsertedObservationsAction, audit.Unsuccessful, auditParams); auditErr != nil {
-			err = auditErr
-		}
-		handleInstanceErr(ctx, err, w, logData)
-		return
-	}
-
-	s.Auditor.Record(ctx, UpdateInsertedObservationsAction, audit.Successful, auditParams)
-
-	log.InfoCtx(ctx, "update imported observations: request successful", logData)
-}
-
-// UpdateImportTask updates any task in the request body against an instance
-func (s *Store) UpdateImportTask(w http.ResponseWriter, r *http.Request) {
-
-	defer request.DrainBody(r)
-
-	ctx := r.Context()
-	vars := mux.Vars(r)
-	instanceID := vars["instance_id"]
-	auditParams := common.Params{"instance_id": instanceID}
-	logData := audit.ToLogData(auditParams)
-	defer r.Body.Close()
-
-	updateErr := func() *taskError {
-		tasks, err := unmarshalImportTasks(r.Body)
-		if err != nil {
-			log.ErrorCtx(ctx, errors.WithMessage(err, "failed to unmarshal request body to UpdateImportTasks model"), logData)
-			return &taskError{err, http.StatusBadRequest}
-		}
-
-		validationErrs := make([]error, 0)
-		var hasImportTasks bool
-
-		if tasks.ImportObservations != nil {
-			hasImportTasks = true
-			if tasks.ImportObservations.State != "" {
-				if tasks.ImportObservations.State != models.CompletedState {
-					validationErrs = append(validationErrs, fmt.Errorf("bad request - invalid task state value for import observations: %v", tasks.ImportObservations.State))
-				} else {
-					if err := s.UpdateImportObservationsTaskState(instanceID, tasks.ImportObservations.State); err != nil {
-						log.ErrorCtx(ctx, errors.WithMessage(err, "Failed to update import observations task state"), logData)
-						return &taskError{err, http.StatusInternalServerError}
-					}
-				}
-			} else {
-				validationErrs = append(validationErrs, errors.New("bad request - invalid import observation task, must include state"))
-			}
-		}
-
-		if tasks.BuildHierarchyTasks != nil {
-			hasImportTasks = true
-			var hasHierarchyImportTask bool
-			for _, task := range tasks.BuildHierarchyTasks {
-				hasHierarchyImportTask = true
-				if err := models.ValidateImportTask(task.GenericTaskDetails); err != nil {
-					validationErrs = append(validationErrs, err)
-				} else {
-					if err := s.UpdateBuildHierarchyTaskState(instanceID, task.DimensionName, task.State); err != nil {
-						if err.Error() == "not found" {
-							notFoundErr := task.DimensionName + " hierarchy import task does not exist"
-							log.ErrorCtx(ctx, errors.WithMessage(err, notFoundErr), logData)
-							return &taskError{errors.New(notFoundErr), http.StatusNotFound}
-						}
-						log.ErrorCtx(ctx, errors.WithMessage(err, "failed to update build hierarchy task state"), logData)
-						return &taskError{err, http.StatusInternalServerError}
-					}
-				}
-			}
-			if !hasHierarchyImportTask {
-				validationErrs = append(validationErrs, errors.New("bad request - missing hierarchy task"))
-			}
-		}
-
-		if tasks.BuildSearchIndexTasks != nil {
-			hasImportTasks = true
-			var hasSearchIndexImportTask bool
-			for _, task := range tasks.BuildSearchIndexTasks {
-				hasSearchIndexImportTask = true
-				if err := models.ValidateImportTask(task.GenericTaskDetails); err != nil {
-					validationErrs = append(validationErrs, err)
-				} else {
-					if err := s.UpdateBuildSearchTaskState(instanceID, task.DimensionName, task.State); err != nil {
-						if err.Error() == "not found" {
-							notFoundErr := task.DimensionName + " search index import task does not exist"
-							log.ErrorCtx(ctx, errors.WithMessage(err, notFoundErr), logData)
-							return &taskError{errors.New(notFoundErr), http.StatusNotFound}
-						}
-						log.ErrorCtx(ctx, errors.WithMessage(err, "failed to update build hierarchy task state"), logData)
-						return &taskError{err, http.StatusInternalServerError}
-					}
-				}
-			}
-			if !hasSearchIndexImportTask {
-				validationErrs = append(validationErrs, errors.New("bad request - missing search index task"))
-			}
-		}
-
-		if !hasImportTasks {
-			validationErrs = append(validationErrs, errors.New("bad request - request body does not contain any import tasks"))
-		}
-
-		if len(validationErrs) > 0 {
-			for _, err := range validationErrs {
-				log.ErrorCtx(ctx, errors.WithMessage(err, "validation error"), logData)
-			}
-			// todo: add all validation errors to the response
-			return &taskError{validationErrs[0], http.StatusBadRequest}
-		}
-		return nil
-	}()
-
-	if updateErr != nil {
-		if auditErr := s.Auditor.Record(ctx, UpdateImportTasksAction, audit.Unsuccessful, auditParams); auditErr != nil {
-			updateErr = &taskError{errs.ErrInternalServer, http.StatusInternalServerError}
-		}
-		log.ErrorCtx(ctx, errors.WithMessage(updateErr, "updateImportTask endpoint: request unsuccessful"), logData)
-		http.Error(w, updateErr.Error(), updateErr.status)
-		return
-	}
-
-	if auditErr := s.Auditor.Record(ctx, UpdateImportTasksAction, audit.Successful, auditParams); auditErr != nil {
-		return
-	}
-
-	log.InfoCtx(ctx, "updateImportTask endpoint: request successful", logData)
-}
-
-func unmarshalImportTasks(reader io.Reader) (*models.InstanceImportTasks, error) {
-
-	b, err := ioutil.ReadAll(reader)
-	if err != nil {
-		return nil, errs.ErrUnableToReadMessage
-	}
-
-	var tasks models.InstanceImportTasks
-	err = json.Unmarshal(b, &tasks)
-	if err != nil {
-		return nil, errs.ErrUnableToParseJSON
-	}
-
-	return &tasks, nil
 }
 
 func unmarshalInstance(ctx context.Context, reader io.Reader, post bool) (*models.Instance, error) {
